Fix some copies of mutexes
Summary:
This patch suppresses automatic generation of the copy-constructor and
assigment operator for the mutex classes, in order to detect at compile
time incorrect copies of these structures (especially if they are
embedded in other classes). It then fixes a couple of instances where
they were indeed copied.

The patch also modifies the list of LightProcesses to be held in an
array instead of a vector. Since a vector can do copies of its elements
for some operations, and LightProcess contains a Mutex, the code was
not compilable anymore. In practice this vector is pretty much used as
a simple array in this case, so it was probably OK already, although
not very safe.

Test Plan:
make fast_tests
make slow_tests

Reviewed By: mwilliams
Reviewers: qixin, myang, mwilliams
CC: macvicar, mwilliams
Revert Plan:
Ok

Differential Revision: 227031