[Fix] methodCall et al need to check their CObjRef
Summary:
A CObjRef doesnt necessarily contain a non-null ObjectData*, so we need to
check it explicitly.

This test case used to seg fault:

<?php
class c {
  function foo() {}
}
class d  {
  function foo() {}
}
function test($a) {
  if ($a == 1) {
    $x = new C;
  } else if ($a == 2) {
    $x = new D;
  }
  $x->foo();
}
test(3);

There doesnt seem to be a good way to add this to our tests, unfortunately.

Also noticed that methodCallEx only has a CObjRef flavor, but its almost always
called with "this" as the first param; so added a ObjectData * overload to avoid
constructing/destructing the Object.

Test Plan:
fast_tests slow_tests.

Reviewed By: qixin
Reviewers: qixin
CC: qixin
Revert Plan:
OK

Differential Revision: 220519