[Yield] Implement Continuation::send()
Summary:
Add a new send() method to Continuation for passing data back into the
generator. Two kinds of yield statements are also added:

$var = yield ...;
list($v1, $v2, ...) = yield ...;

Right now, "yield ..." is not an expression in the general sense, although it
may have a value. It doesn't support composing multiple yields in one
expression.

In the caller, one can call $c->send(123) to send 123 to the generator, and
also to advance the generator to the next step. The first call to the
continuation has to be next(), not send().

A new testcase is also added.

Test Plan:
TestYield

Reviewed By: mwilliams
Reviewers: sgrimm, ahupp, mwilliams, zrait
CC: mwilliams
Tasks:

Revert Plan:
OK

Differential Revision: 220829