Remove setArgPop() and some related code
Summary:
The function setArgPop that I added made the logic more complex. To make the
code simpler, I removed it and made the change so that m_argc is only/always
incremented whenever the argument is pushed. In this way m_argc is always
sync'ed with # of arguments pushed onto the stack. I also removed some other
code that is not used and one argument in the FuncScopeVariableEnvironment
constructor (m_argc always initialized to 0).

Test Plan:
make fast_tests
make slow_tests
run hphpi sandbox and browse the site for sanity check.

Reviewed By: mwilliams
Reviewers: qixin, mwilliams
CC: mwilliams, hphp-diffs@lists
Revert Plan:
OK

Differential Revision: 221344