Fix more issues in the build
Summary:
When SHARED is set, we turn off JEMALLOC and TCMALLOC. But SHARED is only set
when we link the shared library; the files have already been compiled with
-DUSE_JEMALLOC etc. Fix the compilation step to skip the MALLOC related defines
for the .pic.o object files.

Also, ensure that jemalloc and tcmalloc are actually turned off, by moving the
code which does that outside of some debugging macros.

Finally, if OUTDIR_BY_TYPE is turned on, force the shared .so to go to the same
directory it would have gone to if we hadnt turned off jemalloc and tcmalloc;
this is because the .pic.o files have already been built there, and because it
would be too confusing to try and find it otherwise.

Test Plan:
make fast_tests slow_tests

Reviewed By: qixin
Reviewers: asharma, qixin
CC: achao, qixin
Revert Plan:
OK

Differential Revision: 223095