Update inconsistencies doc
Summary:
A while ago we decided to diverge slightly from Zend semantics when a copy
of an array is made to honor copy-on-write semantics. This decision was
made because it is basically impossible for us to perfectly mimic cases
where Zend violates copy-on-write semantics due to static arrays and
SharedMap.

This diff updates the inconsistencies doc to describe how we deal with the
"NextFreeElement" field when a copy of an array is made.

Test Plan:
Inspection.

Reviewed By: je
Reviewers: qixin, je, myang, mwilliams
CC: je
Revert Plan:
OK

Differential Revision: 228044