[Fix/Perf] Variant::bindNoVariant() without FAST_REF_COUNT_FOR_VARIANT
Summary:
This method was declared with FAST_REF_COUNT_FOR_VARIANT, but might be used
without it. This diff fixes the problem by implementing the no-fast-ref-count
case for bindNoVariant(). It also fixes a problem for Variant::destruct()
without the macro.

Also, a Variant cannot have its m_data be null without KindOfUninit or
KindOfNull, so it is not necessary to check for null if the source Variant is of
type KindOfString, KindOfArray, KindOfObject, or KindOfVariant.

Test Plan:
build, with and without FAST_REF_COUNT_FOR_VARIANT
fast_tests

Reviewed By: macvicar
Reviewers: macvicar, myang
CC: macvicar
Revert Plan:
OK

Differential Revision: 229394