Leak in evhttp_async_post()
Summary:
The data isn't swept so at the end of the request the client and file
descriptor remains open.

Mark the resource as sweepable so it doesn't leak.

Added printf to the destructor and checked it was called
<?php
$post_data = '';
$host = 'httpstat.us';
$headers = array("Host: $host","Content-Type:
application/x-www-form-urlencoded","Connection: close","Content-Length:
".strlen($post_data));
$task = evhttp_async_post("http://httpstat.us/200",$post_data, $headers, 1);
$result = evhttp_recv($task);
var_dump($result);

Reviewed By: mwilliams
Reviewers: myang, mwilliams
CC: mwilliams
Revert Plan:
Ok

Differential Revision: 233021