Add sandbox support to PageletServer
Summary:
PageletServer isn't aware of sandbox and it always uses /var/www because
that's SourceRootInfo::path() returns when sandbox is off (it returns
RuntimeOption::SourceRoot which is the value of Server.SourceRoot. Therefore
the PageletServer does not access user's intended file (unless a symlink is
set to user's sandbox www). I made a change in f_pagelet_server_task_start so
that SandboxMode is checked and if the caller has not explicitly passed in
the header, we will pass on the "Host" header from the original thread.

Task ID: #

Blame Rev:

Reviewers:
qigao, mwilliams
CC:
hphp-diffs@lists.facebook.com
Revert Plan:

Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 233402