[Fix] Mark the class as volatile based on its name usage
Summary:
Our current mark-as-volatile method doesn't handle the
case described in test plan very well. This diff fixes that problem by
marking the simple variable in symbol table and use that to tell whether
the literal strings in switch are likely to be class names. If they are,
mark these classes as volatile since they are likely loaded dynamically.

Differential Revision: 233739