[Fix] Fix crash in unserialize
Summary:
Another case of a bad key being passed to lvalAt. This time, its because we
explicitly tell lvalAt that the key is good, so it doesnt bother to call toKey
on it. We cant call toKey, because eg "1" should be different from 1 in this
specific instance - where toKey would treat them as the same thing.

So we need to check explicitly for integer or string keys, throw if not, and
then tell lvalAt to use the key as is.

Reviewed By: myang
Reviewers: andrewparoski, myang
CC: myang, lesha
Tasks:

Differential Revision: 235453