[Perf/Fix] Better handling of reference parameters
Summary:
This is part of the ongoing effort to get rid of the contagious flag. Reference
params are now represented by VRefParam (const VRefParamValue &), rather than
variant. There is an automatic conversion from VRefParam to Variant&, and
passing a Variant or CVarRef where a VRefParam is expected will automatically
create a temporary Variant (just as it would have before). ref has been modified
to return a RefReturn, which when passed to a VRefParam will construct a
strongly bound Variant without using the contagious flag. When a RefResult is
converted to a CVarRef, it will set the contagious flag, just like before (so
existing, non generated code mostly works without change; a few things fail to
compile, but are easy to fix; if it compiles, it should work).

The advantage is that now VRefParamValue's can be constructed early, and passed
through multiple levels of wrapper untouched; rather than trying to pass a
CVarRef through multiple levels, and having to construct Variants to avoid
aliasing artifacts.

I regenerated all of ext from the idl (some changes to base.idl were needed),
and fixed up parts of the runtime to use the more efficient ways of doing
things. I noticed that g++ was no longer doing a good job of predicting the
common path in the invoke helpers, so added hints to help it. The overall code
size went down, and the common path became significantly shorter.

This also fixes some issues with exactly when the contagious bit is read. Test
case included

Test Plan:
fast_tests slow_tests

Reviewed By: myang
Reviewers: qigao, myang
Commenters: qigao
CC: stephentu, mpal, bill, cbueno, minwenji, ps, mwilliams, myang, qigao
Differential Revision: 246842