[Fix] Fix ref params in constructor with type hint
Summary:
For a constructor with a type hint, the create function's declaration was
generated with CVarRefs (intended to be pass-through), but the body of the
function called ref() on the reference parameters. This is not needed, since
create's caller will have done so, and wrong in the case that a shared variant
was passed in (in particular, a null default argument).

Test Plan:
fast_tests slow_tests

Reviewed By: myang
Reviewers: je, qigao, myang
CC: ps, mwilliams, myang
Differential Revision: 241435
Tasks: # 546002 Error: Fatal PHP Exception: UnsatisfiedDependencyException