[Fix] Calling parent::__construct should find a class name constructor
Summary:
This worked as long as the parent class was not redeclared. But in that case,
we did a dynamic lookup, just for the name __construct.

Also fix an issue with runtime/tmp/run.sh. It contained the value of $HPHP_LIB
at the time run.sh was created, rather than the variable $HPHP_LIB. Since we
only create the file if it doesnt exist, this meant we could be using a stale
version of libhphpruntime.so

Test Plan:
fast_tests slow_tests (new test case added).

Reviewed By: qigao
Reviewers: qigao
CC: bhu, ps, qigao
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 243895