[Fix] Fix HPHPi bug with yield and closures
Summary:
We do syntactic transformation of a function containing yield statements at
the AST level. The original function body is turned into a statement that
returns a Continuation object that is associated with a internally created
closure function. In HPHPi, the associated closure function must be declared
before it is used. In the newly added test case, the associated closure
function isn't declared when it is used. I renamed the pendingStatements
to prependingStatements and prepend the closure function before the original
statement from which the closure function is spinned off so the closure
function is declared prior to its usage.
Task ID: # 554932

Blame Rev:

Reviewers:
qigao, mwilliams
CC:
hphp-diffs@lists
Test Plan:
make fast_tests
make slow_tests
Revert Plan:

Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 245798