[Fix] Specialized int switch case is now properly generated with respect to PHP switch semantics
Summary:
HPHP tries to specialize switch statements in the case of all int
literals, however it does so in a way that is incorrect with regards to
PHP semantics by blindly converting the object into an int and then
switching on the converted int.

This fix addresses this issue by introducing a new Variant method
called hashForIntSwitch(), which properly computes a switch "hash"
for the specific data type. The compiler statically computes two
key hashes, one which is used to switch into the first non-zero
case, and one which is used to switch into the default case.

As a result, we are still able to optimize switch statements
with only integer constant cases as before, except in the case
where the programmer enumerates all 2^64 integer cases. In this
*rare* case, we simply fall back to the un-optimized version.

Reviewed By: mwilliams
Reviewers: mwilliams
CC: ps, mwilliams
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 246711