Fix www-dev build
Summary:
Up the buffer sizes from 32kb to 128kb and 64kb respetively

We had a single file that was on the 32kb boundary that split a token.

Going to look into a fix in flex for this so mmap can be used.

Task ID: #

Blame Rev:

Reviewers: oyamauchi

CC:

Test Plan:
build www
build www-dev

Revert Plan:
Ok

Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 249201