[Perf] Remove static method wrappers where possible
Summary:
Currently, every static method gets the class name passed in, which it uses to
implement per-class static variables. If the method doesnt contain any static
variables, or the method's class is not derived from, or the method is private,
we dont need this parameter, so remove it.

Test Plan:
fast_tests slow_tests

Reviewed By: qigao
Reviewers: qigao, myang
CC: ps, mwilliams, qigao
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 254721