[Perf/Fix] Cleanup object creation
Summary:
When we call X::foo() on a non static method, with no $this available, we
synthesize a dummy object. The X::createDummy method was created for this
purpose. However, it called init(), which sets up all the explicitly declared
properties on the object, which is wrong, because it can trigger __autoload,
and/or generate missing class errors.

Simple fix would be to remove the call to init. But I also noticed that we have
a function coo_X, which creates an object, and calls init on it. The coo_*
functions were only used from create_object_only, so its actually better (for
code size) to change them to /not/ call init, and call init from
create_object_only instead. Once thats done, we can re-use the coo_X functions
to create the dummy objects, and get rid of createDummy altogether.

createDummy was only used from the X::i_* and X::ifa_* dynamic dispatch
helpers. The vast majority of calls to these dont need to create an object, but
the body looked something like this:

  c_X *self;
  p_X obj;
  if (mcp.obj) {
     self = mcp.obj;
  } else {
    obj = createDummy();
    self = obj.get();
  }

This meant that on every call, we created, and then destroyed obj. This is
relatively cheap, but uses a surprising amount of code. In addition, it causes
an unwind handler to be generated, causing even more code bloat.

So I created i_dummy, and ifa_dummy routines that take care of the object
creation and destruction. This reduces the size of the dynamic dispatch helpers
dramatically, and speeds them up a lot in the case where no object was
necessary.

I then noticed that we could reuse coo_X in the implementation of cloneImpl,
creating another big code size reduction.

While looking at the generated assembly code, I noticed X::sweep methods for
every class, which just called the destructor. Since the destructor is virtual,
we dont need to generate one for every class. It turns out they werent used
anyway.

Finally, used coo_X() rather than NEWOBJ(c_X)() for creating objects in new
expressions. Another big code size reduction.

I noticed an inefficient Object constructor, which default initialized the
SmartPtr, and then called the assignment operator. I initially rewrote it to
directly initialize the SmartPtr, but then realized it was better to just remove
it. Another small code size reduction.

Test Plan:
fast_tests slow_tests

Reviewed By: myang
Reviewers: myang, qigao
CC: cbueno, ps, mwilliams, myang, mpal
Revert Plan:
Tags:

- begin *PUBLIC* platform impact section -
Bugzilla: #
- end platform impact -

Differential Revision: 256860